[Feat] 방 데이터 구조 변경 - Room 클래스 내에 게임오브젝트를 관리하는 RoomEntity와 방의 데이터를 가지고 있는 RoomData가 있었으나, 현재로서는 큰 필요성을 느끼지 못하여 Room에서 두 기능을 모두 수행하는 것으로 변경
