--- Assets/Scripts/Room/RoomData.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 92366bda169733142b5e7252a3e7213b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
