--- Assets/Scripts/Room/RoomEntity.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c023bf44adf2b1e4ba25ace592c0e2f7
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
